smart_research_backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.2
sentence-transformers==2.3.1
groq==0.4.2
pydantic==2.5.3
supabase==2.25.0
